FASTAPI/requirements.txt
Pillow
numpy
uvicorn
fastapi
tensorflow
prometheus_client
starlette
h5py